cmd: implement tls termination capability
closes #177
