+++ cmd/server/handler.go
@@ -14,9 +14,60 @@ import (
 	"github.com/ory-am/hydra/policy"
 	"github.com/ory-am/hydra/warden"
 	"github.com/ory-am/ladon"
-	"os"
-	"strings"
+	"time"
+	"crypto/tls"
+	"github.com/urfave/negroni"
+	"github.com/meatballhat/negroni-logrus"
+	"github.com/ory-am/hydra/herodot"
+	"net/http"
+
+	"github.com/spf13/cobra"
+	"golang.org/x/net/context"
 )
+
+func RunHost(c *config.Config) func(cmd *cobra.Command, args []string) {
+	return func(cmd *cobra.Command, args []string) {
+		router := httprouter.New()
+		serverHandler := &Handler{Config: c}
+		serverHandler.start(router)
+
+		if ok, _ := cmd.Flags().GetBool("dangerous-auto-logon"); ok {
+			logrus.Warnln("Do not use flag --dangerous-auto-logon in production.")
+			err := c.Persist()
+			pkg.Must(err, "Could not write configuration file: %s", err)
+		}
+
+		n := negroni.New()
+		n.Use(negronilogrus.NewMiddleware())
+		n.UseFunc(serverHandler.rejectInsecureRequests)
+		n.UseHandler(router)
+		http.Handle("/", n)
+
+		var srv = http.Server{
+			Addr: c.GetAddress(),
+			TLSConfig: &tls.Config{
+				Certificates: []tls.Certificate{
+					getOrCreateTLSCertificate(cmd, c),
+				},
+			},
+			ReadTimeout: time.Second * 5,
+			WriteTimeout: time.Second * 10,
+		}
+
+		var err error
+		logrus.Infof("Setting up http server on %s", c.GetAddress())
+		if ok, _ := cmd.Flags().GetBool("force-dangerous-http"); ok {
+			logrus.Warnln("HTTPS disabled. Never do this in production.")
+			err = srv.ListenAndServe()
+		} else if (c.AllowTLSTermination != "") {
+			logrus.Infoln("TLS termination enabled, disabling https.")
+			err = srv.ListenAndServe()
+		} else {
+			err = srv.ListenAndServeTLS("", "")
+		}
+		pkg.Must(err, "Could not start server: %s %s.", err)
+	}
+}
 
 type Handler struct {
 	Clients     *client.Handler
@@ -25,9 +76,11 @@ type Handler struct {
 	OAuth2      *oauth2.Handler
 	Policy      *policy.Handler
 	Warden      *warden.WardenHandler
+	Config      *config.Config
 }
 
-func (h *Handler) Start(c *config.Config, router *httprouter.Router) {
+func (h *Handler) start(router *httprouter.Router) {
+	c := h.Config
 	ctx := c.Context()
 
 	// Set up warden
@@ -60,77 +113,19 @@ func (h *Handler) Start(c *config.Config, router *httprouter.Router) {
 	h.createRootIfNewInstall(c)
 }
 
-func (h *Handler) createRS256KeysIfNotExist(c *config.Config, set, lookup string) {
-	ctx := c.Context()
-	generator := jwk.RS256Generator{}
-
-	if _, err := ctx.KeyManager.GetKey(set, lookup); errors.Is(err, pkg.ErrNotFound) {
-		logrus.Infof("Key pair for signing %s is missing. Creating new one.", set)
-
-		keys, err := generator.Generate("")
-		pkg.Must(err, "Could not generate %s key: %s", set, err)
-
-		err = ctx.KeyManager.AddKeySet(set, keys)
-		pkg.Must(err, "Could not persist %s key: %s", set, err)
-	}
-}
-
-func (h *Handler) createRootIfNewInstall(c *config.Config) {
-	ctx := c.Context()
-
-	clients, err := h.Clients.Manager.GetClients()
-	pkg.Must(err, "Could not fetch client list: %s", err)
-	if len(clients) != 0 {
+func (h *Handler) rejectInsecureRequests(rw http.ResponseWriter, r *http.Request, next http.HandlerFunc) {
+	if r.TLS != nil || h.Config.ForceHTTP {
+		next.ServeHTTP(rw, r)
 		return
 	}
 
-	rs, err := pkg.GenerateSecret(16)
-	pkg.Must(err, "Could notgenerate secret because %s", err)
-	secret := string(rs)
-
-	id := ""
-	forceRoot := os.Getenv("FORCE_ROOT_CLIENT_CREDENTIALS")
-	if forceRoot != "" {
-		credentials := strings.Split(forceRoot, ":")
-		if len(credentials) == 2 {
-			id = credentials[0]
-			secret = credentials[1]
-		} else {
-			logrus.Warnln("You passed malformed root client credentials, falling back to random values.")
-		}
+	if err := h.Config.DoesRequestSatisfyTermination(r); err == nil {
+		next.ServeHTTP(rw, r)
+		return
+	} else {
+		logrus.WithError(err).Warnln("Could not serve http connection")
 	}
 
-	logrus.Warn("No clients were found. Creating a temporary root client...")
-	root := &client.Client{
-		ID: id,
-		Name:          "This temporary client is generated by hydra and is granted all of hydra's administrative privileges. It must be removed when everything is set up.",
-		ResponseTypes: []string{"id_token", "code", "token"},
-		GrantTypes:    []string{"implicit", "refresh_token", "authorization_code", "password", "client_credentials"},
-		GrantedScopes: []string{"hydra", "core", "openid", "offline"},
-		RedirectURIs:  []string{"http://localhost:4445/callback"},
-		Secret:        secret,
-	}
-
-	err = h.Clients.Manager.CreateClient(root)
-	pkg.Must(err, "Could not create temporary root because %s", err)
-	err = ctx.LadonManager.Create(&ladon.DefaultPolicy{
-		Description: "This is a policy created by hydra and issued to the first client. It grants all of hydra's administrative privileges to the client and enables the client_credentials response type.",
-		Subjects:    []string{root.GetID()},
-		Effect:      ladon.AllowAccess,
-		Resources:   []string{"rn:hydra:<.*>"},
-		Actions:     []string{"<.*>"},
-	})
-	pkg.Must(err, "Could not create admin policy because %s", err)
-
-	c.Lock()
-	c.ClientID = root.ID
-	c.ClientSecret = string(secret)
-	c.Unlock()
-
-	logrus.Infoln("Temporary root client created.")
-	if forceRoot == "" {
-		logrus.Infoln("client_id: %s", root.GetID())
-		logrus.Infoln("client_secret: %s", string(secret))
-		logrus.Warn("WARNING: YOU MUST delete this client once in production, as credentials may have been leaked logfiles.")
-	}
+	ans := new(herodot.JSON)
+	ans.WriteErrorCode(context.Background(), rw, r, http.StatusBadGateway, errors.New("Can not serve request over insecure http"))
 }
